fix: required prop fixes.
